remove proxy manager
