+++ nmproxy/nm-proxy.go
@@ -8,20 +8,15 @@ import (
 	ncconfig "github.com/gravitl/netclient/config"
 	"github.com/gravitl/netclient/ncutils"
 	"github.com/gravitl/netclient/nmproxy/config"
-	"github.com/gravitl/netclient/nmproxy/manager"
-	ncmodels "github.com/gravitl/netclient/nmproxy/models"
 	"github.com/gravitl/netclient/nmproxy/turn"
 	"github.com/gravitl/netclient/nmproxy/wg"
 	"github.com/gravitl/netmaker/logger"
-	"github.com/gravitl/netmaker/models"
 )
 
 // Start - setups the global cfg for proxy and starts the proxy server
-func Start(ctx context.Context, waitg *sync.WaitGroup,
-	mgmChan chan *models.HostPeerUpdate, hostNatInfo *ncmodels.HostInfo) {
+func Start(ctx context.Context, waitg *sync.WaitGroup) {
 
 	defer waitg.Done()
-
 	config.InitializeCfg()
 	defer config.Reset()
 	wgIface, err := wg.GetWgIface(ncutils.GetInterfaceName())
@@ -31,8 +26,6 @@ func Start(ctx context.Context, waitg *sync.WaitGroup,
 	}
 	config.GetCfg().SetIface(wgIface)
 	proxyWaitG := &sync.WaitGroup{}
-	proxyWaitG.Add(1)
-	go manager.Start(ctx, proxyWaitG, mgmChan)
 	proxyWaitG.Add(1)
 	go turn.WatchPeerSignals(ctx, proxyWaitG)
 	turnCfgs := ncconfig.GetAllTurnConfigs()
